Move error related files into features folder
